whatsapp_reminder/src/requirements.txt
asgiref==3.8.1
Django==4.2.19
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
psycopg2==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing-extensions==4.12.2
requests==2.32.3
gunicorn==23.0.0
drf-spectacular==0.28.0
drf-spectacular-sidecar